Refactor tournaments adapter: replace ITournamentsAdapter and TournamentsAdapter with IGetTournamentsAdapter and GetTournamentsAdapter, updating method names and restructuring tests accordingly.
