+++ src/BowlingMegabucks.TournamentManager.Presentation/Tournaments/GetTournaments/IGetTournamentsAdapter.cs
@@ -1,0 +1,9 @@
+using BowlingMegabucks.TournamentManager.Presentation.Tournaments.GetTournaments;
+using ErrorOr;
+
+namespace BowlingMegabucks.TournamentManager.Presentation.Tournaments.GetTournaments;
+
+internal interface IGetTournamentsAdapter
+{
+    Task<ErrorOr<OffsetPagingResult<TournamentSummaryViewModel>>> ExecuteAsync(int? page, int? pageSize, CancellationToken cancellationToken);
+}
