--- src/BowlingMegabucks.TournamentManager.Presentation/Tournaments/ITournamentsAdapter.cs
@@ -1,9 +1,0 @@
-using BowlingMegabucks.TournamentManager.Presentation.Tournaments.GetTournaments;
-using ErrorOr;
-
-namespace BowlingMegabucks.TournamentManager.Presentation.Tournaments;
-
-internal interface ITournamentsAdapter
-{
-    Task<ErrorOr<OffsetPagingResult<TournamentSummaryViewModel>>> GetTournamentsAsync(int? page, int? pageSize, CancellationToken cancellationToken);
-}
